+++ api_types.ts
@@ -721,7 +721,7 @@ export interface TemporaryBlobStorage {
   storeUrl(
     url: string,
     opts?: {expiryMs?: number; downloadFilename?: string; contentType?: string},
-    fetchOpts?: Pick<FetchRequest, 'disableAuthentication' | 'headers'>,
+    fetchOpts?: Pick<FetchRequest, 'disableAuthentication' | 'headers' | 'cacheTtlSecs'>,
   ): Promise<string>;
   /**
    * Stores the given data as a file with the given content type in Coda-hosted temporary storage.